New CORS Proxy + Ignore HTTPS for pond hockey
Ok, so these two ideas would be VERY BAD for 99% of applications. We are actively going out of our way to ignore a lot of good security that we don't need for this application. Unless you are 100% sure that you don't need these things, don't do what we're doing here.

1) We were using a CORS proxy to get around CORS header blocks in the browser without updating the server. I should talk to the snoking admins to see if I can't get them to just add CORS in.
2) Pond Hockey's cert is invalid. They should fix this, but it's all public data so unless someone wants to start serving fake pond schedules, it's fine